+++ python/src/validations/validator.py
@@ -1,4 +1,4 @@
-from exceptions.value_not_number_error import ValueNotNumberError 
+from src.exceptions.value_not_number_error import ValueNotNumberError 
     
 def validate_number(number):
